MCP/server/py-calculator/pyproject.toml
[project]
requires-python = ">=3.13"
name = "calculator"
version = "0.1.0"
description = "A calculator MCP server with basic arithmetic operations"
authors = []
dependencies = [
    "mcp @ git+https://github.com/modelcontextprotocol/python-sdk.git",
    "uvicorn",
    "python-dotenv>=1.0.0",
    "arize-phoenix>=4.0.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "openinference-instrumentation>=0.1.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]